Обновление ДЗ: калькулятор, Koin, MVVM
