+++ app/build.gradle.kts
@@ -61,4 +61,6 @@ dependencies {
     implementation("androidx.core:core-ktx:1.13.1")
     implementation("androidx.appcompat:appcompat:1.7.0")
     implementation("androidx.lifecycle:lifecycle-runtime-ktx:2.8.4")
+    implementation("net.objecthunter:exp4j:0.4.8")
+    implementation("io.insert-koin:koin-android:3.5.6")
 }
