+++ app/src/main/AndroidManifest.xml
@@ -4,6 +4,7 @@
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
+        android:name=".App"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
@@ -15,27 +16,27 @@
         tools:targetApi="31">
 
         <activity
-            android:name=".ui.WeatherDetailActivity"
-            android:parentActivityName=".ui.WeatherActivity" />
+            android:name=".feature.weather.presentation.ui.WeatherDetailActivity"
+            android:parentActivityName=".feature.weather.presentation.ui.WeatherActivity" />
 
         <activity
-            android:name=".ui.WeatherActivity"
+            android:name=".feature.weather.presentation.ui.WeatherActivity"
             android:exported="false" />
 
         <activity
-            android:name=".ListActivity"
+            android:name=".feature.cars.presentation.ui.ListActivity"
             android:exported="false" />
 
         <activity
-            android:name=".CalculatorActivity"
+            android:name=".feature.calculator.presentation.ui.CalculatorActivity"
             android:exported="false" />
 
         <activity
-            android:name=".ui.HomeActivity"
+            android:name=".feature.home.presentation.ui.HomeActivity"
             android:exported="false" />
 
         <activity
-            android:name=".MainActivity"
+            android:name=".feature.home.presentation.ui.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
