+++ app/src/main/java/com/example/topacademy_android/di/coreModule.kt
@@ -1,0 +1,10 @@
+package com.example.topacademy_android.di
+
+
+import org.koin.dsl.module
+
+
+val coreModule = module {
+
+
+}
