--- app/src/main/java/com/example/topacademy_android/network/RetrofitClient.kt
@@ -1,16 +1,0 @@
-package com.example.topacademy_android.network
-
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-object RetrofitClient {
-    private const val BASE_URL = "https://www.7timer.info/"
-
-    val api: WeatherApi by lazy {
-        Retrofit.Builder()
-            .baseUrl(BASE_URL)
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-            .create(WeatherApi::class.java)
-    }
-}
