+++ app/src/main/res/layout/activity_calculator.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CalculatorActivity">
+    tools:context=".feature.calculator.presentation.ui.CalculatorActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar"
