--- app/src/main/res/layout/activity_main.xml
@@ -1,131 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/background_bottom"
-    tools:context=".MainActivity">
-
-    <View
-        android:id="@+id/topShape"
-        android:layout_width="@dimen/_0dp"
-        android:layout_height="@dimen/_0dp"
-        android:background="@drawable/top_shape"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintBottom_toTopOf="@+id/loginCard" />
-
-    <androidx.cardview.widget.CardView
-        android:id="@+id/loginCard"
-        android:layout_width="@dimen/_0dp"
-        android:layout_height="wrap_content"
-        android:layout_margin="@dimen/_24dp"
-        app:cardBackgroundColor="@color/background_bottom"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toTopOf="@+id/startButton"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent">
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:gravity="center_horizontal"
-            android:orientation="vertical"
-            android:padding="@dimen/_24dp">
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="@string/the_tolet"
-                android:textColor="@color/text_main"
-                android:textSize="@dimen/_24sp"
-                android:textStyle="bold" />
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginBottom="@dimen/_24dp"
-                android:text="@string/find_your_house"
-                android:textColor="@color/text_main"
-                android:textSize="@dimen/_14sp" />
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginBottom="@dimen/_16dp"
-                android:text="@string/login"
-                android:textColor="@color/text_main"
-                android:textSize="20sp"
-                android:textStyle="bold" />
-
-            <EditText
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:minHeight="@dimen/_48dp"
-                android:hint="@string/email"
-                android:inputType="textEmailAddress"
-                android:textColor="@color/text_main"
-                android:textColorHint="@color/text_main"
-                android:autofillHints="emailAddress" />
-
-            <EditText
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="@dimen/_8dp"
-                android:minHeight="@dimen/_48dp"
-                android:hint="@string/password"
-                android:inputType="textPassword"
-                android:textColor="@color/text_main"
-                android:textColorHint="@color/text_main"
-                android:autofillHints="password" />
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="end"
-                android:layout_marginTop="@dimen/_4dp"
-                android:text="@string/forgot"
-                android:textColor="@color/text_main" />
-
-            <Button
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="@dimen/_16dp"
-                android:backgroundTint="@color/button_bg2"
-                android:text="@string/log_in"
-                android:textColor="@color/button_text3" />
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="@dimen/_16dp"
-                android:text="@string/or_continue_with"
-                android:textColor="@color/text_main" />
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="@dimen/_16dp"
-                android:text="@string/don_t_have_account_create_now"
-                android:textColor="@color/text_main" />
-
-        </LinearLayout>
-    </androidx.cardview.widget.CardView>
-
-    <com.google.android.material.button.MaterialButton
-        android:id="@+id/startButton"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/Begin"
-        android:textColor="@color/button_text3"
-        android:backgroundTint="@color/button_bg"
-        android:layout_marginBottom="@dimen/_32dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
