+++ app/src/main/res/values/dimens.xml
@@ -11,6 +11,7 @@
     <dimen name="search_padding_end">10dp</dimen>
     <dimen name="search_text_size">50sp</dimen>
     <dimen name="_38sp">38sp</dimen>
+    <dimen name="_20sp">20sp</dimen>
     <dimen name="_18sp">18sp</dimen>
     <dimen name="_160dp">160dp</dimen>
     <dimen name="_24dp">24dp</dimen>
